// Frontend/src/app/models/staff.ts
export class Staff {
    staffId: number;
    name: string;
    phone: string;
    email: string;
    gender: string;
    dob: string;
    designation: string;
    dateOfJoining: string;
    department: string;
    experience: string;

    constructor(){
        this.staffId = 0;
        this.name = '';
        this.phone = '';
        this.email = '';
        this.gender = '';
        this.dob = '';
        this.designation = '';
        this.dateOfJoining = '';
        this.department = '';
        this.experience = '';
    }
}